backend/requirements.txt
absl-py==1.4.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
fonttools==4.42.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.57.0
gunicorn==21.2.0
idna==3.4
joblib==1.3.2
keras==3.7.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.24.3
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
tensorflow==2.17.0
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
whitenoise==5.0.1
wrapt==1.15.0
rdkit==2023.3.1
iterative-stratification==0.1.7
pubchempy
